fix: move types to a separate module
